+++ CHANGELOG.md
@@ -8,8 +8,11 @@ Todos los cambios notables del proyecto se documentarán aquí.
 - Páginas HTML (index y proyectos)
 - Estilos base CSS
 - Configuración JavaScript
-
 ## En desarrollo - 2025-11-17
 ### Añadido
 - Tarjeta de presentación de Alexis Sevilla Serrano por @Alexis15014
 - Proyecto AJK por @Alexis15014
+## [En desarrollo] - 2025-11-17
+### Añadido
+- Tarjeta de presentación de Kevin por @keviin17
+- Proyecto AJK por @keviin17